codegen fixes, error conversion

+++ src/serde_xdr/deserializer.rs
@@ -28,7 +28,7 @@ macro_rules! impl_num {
 
 pub struct Deserializer<R: Read> {
     reader: R,
-    bytes_consumed: u32
+    bytes_consumed: usize
 //    first: Option<u8>,
 }
 
@@ -36,12 +36,12 @@ impl<R: Read> Deserializer<R> {
     pub fn new(reader: R) -> Deserializer<R> {
         Deserializer {
             reader: reader,
-            bytes_consumed: 0u32
+            bytes_consumed: 0,
 //            first: None,
         }
     }
 
-   pub fn get_bytes_consumed(&self) -> u32 {
+   pub fn get_bytes_consumed(&self) -> usize {
         self.bytes_consumed
    }
 }
