+++ src/serde_xdr/lib.rs
@@ -23,13 +23,13 @@ pub fn to_bytes<T>(value: &T, buf: &mut Vec<u8>) -> EncoderResult<()>
     Ok(())
 }
 
-pub fn from_reader<T: Deserialize, R: Read>(reader: R) -> DecoderResult<(T, u32)> {
+pub fn from_reader<T: Deserialize, R: Read>(reader: R) -> DecoderResult<(T, usize)> {
     let mut de = Deserializer::new(reader);
     let value = try!(Deserialize::deserialize(&mut de));
 //    try!(de.end());
     Ok((value, de.get_bytes_consumed()))
 }
 
-pub fn from_bytes<T: Deserialize>(v: &[u8]) -> DecoderResult<(T, u32)> {
+pub fn from_bytes<T: Deserialize>(v: &[u8]) -> DecoderResult<(T, usize)> {
     from_reader(v)
 }
